Fix compile error in C99 Modify code structure
